requirements.txt
discord.py>=2.5.2
nltk>=3.9.1
language_tool_python>=2.9.4
profanity_check>=1.0.3
joblib>=1.4.2